requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
SQLAlchemy==2.0.23
mysqlclient==2.2.4
alembic==1.12.1

# Authentication & Security
PyJWT==2.8.0
cryptography==42.0.5
bcrypt==4.1.2

# Redis
redis==5.0.1

# Elasticsearch
elasticsearch==8.13.0

# WebSocket
python-socketio==5.11.1
eventlet==0.33.3

# Production
gunicorn==21.2.0
gevent==23.9.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
bleach==6.1.0

# Development (optional)
pytest==7.4.3
pytest-flask==1.2.0
black==23.11.0
flake8==6.1.0